requirements.txt
flask
sqlalchemy
uvicorn
pyjwt
Flask-WTF
Flask-Bootstrap4
flask_bootstrap
pysam